+++ lib/src/dictosaurus.dart
@@ -13,38 +13,25 @@ import 'package:dictosaurus/src/_index.dart';
 /// - [startsWith] returns terms from a [KGramsMap] that start with a sequence
 ///   of characters; and
 /// - [expandTerm] expands a term using [synonymsOf] and [suggestionsFor].
-abstract class DictoSaurus {
+abstract class DictoSaurus implements Dictionary, Thesaurus, AutoCorrect {
   //
 
   /// Initializes a [DictoSaurus] with [autoCorrect] and [dictionary] instances.
   factory DictoSaurus(
           {required Dictionary dictionary, required AutoCorrect autoCorrect}) =>
-      _DictoSaurusImpl(autoCorrect, dictionary);
-
-  /// Returns the meaning of [term].
-  Future<DictionaryEntry?> getEntry(Term term);
-
-  /// Returns the synonyms of [term] and its tokenized versions.
-  Future<Set<String>> synonymsOf(String term);
-
-  /// Returns a set of unique alternative spellings for a [term].
-  ///
-  /// If [limit] is not null, only the best [limit] matches will be returned.
-  Future<Map<String, List<String>>> suggestionsFor(String term, [int limit]);
-
-  /// Returns a set of unique terms that start with [chars].
-  Future<List<String>> startsWith(String chars);
+      _DictoSaurusImpl(dictionary, autoCorrect);
 
   /// Expands [term] to an ordered list of terms.
   ///
-  /// If [term] exists in the dictionary, the returned list will include
-  /// the [term] and then its synonyms.
+  /// The returned list always has [term] as its first element.
   ///
-  /// If [term] is not found in the dictionary, terms with similar spellings
-  /// will be returned, ordered in descending order of relevance (i.e. best
-  /// match first).
+  /// Any synonyms returned by the dictionary provider are added to the list.
   ///
-  /// If [limit] is not null, only the first [limit] terms will be returned.
+  /// If the length of the list is less than [limit], terms with similar
+  /// spellings will be added, ordered in descending order of relevance (i.e.
+  /// best match first).
+  ///
+  /// Only the first [limit] terms will be returned.
   Future<List<String>> expandTerm(String term, [int limit]);
 }
 
@@ -60,51 +47,74 @@ abstract class DictoSaurusBase with DictoSaurusMixin {
   const DictoSaurusBase();
 }
 
-/// A mixin class that implements [DictoSaurus.getEntry],
-/// [DictoSaurus.synonymsOf], [DictoSaurus.suggestionsFor] and
-/// [DictoSaurus.expandTerm].
+/// A mixin class that implements the [Dictionary], [Thesaurus] and
+/// [AutoCorrect] interface methods as well as [DictoSaurus.expandTerm].
 ///
-/// Implementations that mix in [DictoSaurusMixin] must override:
-/// - [dictionary] is ;
+/// Implementations that mix in [DictoSaurusMixin] must override [dictionary],
+/// [autoCorrect] and [Thesaurus].
 abstract class DictoSaurusMixin implements DictoSaurus {
   //
 
-  /// A [Dictionary] instance used by [getEntry].
-  Dictionary get dictionary;
+  /// A [Thesaurus] instance used by [getEntry].
+  Thesaurus get thesaurus;
 
   /// A [AutoCorrect] instance used by [suggestionsFor].
   AutoCorrect get autoCorrect;
 
+  /// A [Dictionary] instance used by [getEntry].
+  Dictionary get dictionary;
+
+  @override
+  String get languageCode => dictionary.languageCode;
+
+  @override
+  Future<Set<String>> synonymsOf(String term, [PartOfSpeech? partOfSpeech]) =>
+      thesaurus.synonymsOf(term, partOfSpeech);
+
+  @override
+  Future<Set<String>> antonymsOf(String term, [PartOfSpeech? partOfSpeech]) =>
+      thesaurus.synonymsOf(term, partOfSpeech);
+
+  @override
+  Future<Set<String>> inflectionsOf(String term,
+          [PartOfSpeech? partOfSpeech]) =>
+      dictionary.inflectionsOf(term, partOfSpeech);
+
+  @override
+  Future<Set<String>> phrasesWith(String term, [PartOfSpeech? partOfSpeech]) =>
+      dictionary.phrasesWith(term, partOfSpeech);
+
+  @override
+  Future<Set<String>> definitionsFor(String term,
+          [PartOfSpeech? partOfSpeech]) =>
+      dictionary.definitionsFor(term, partOfSpeech);
+
   @override
   Future<List<String>> expandTerm(String term, [int limit = 5]) async {
-    final retVal = <Term>[];
+    final retVal = <String>{term};
     final entry = await dictionary.getEntry(term);
     if (entry != null) {
       retVal.add(term);
-      retVal.addAll(entry.allSynonyms);
-    } else {
-      final suggestions = await autoCorrect.suggestionsFor(term, limit);
-      for (final s in suggestions.values) {
-        retVal.addAll(s);
-      }
+      retVal
+          .addAll(entry.allSynonyms.where((element) => !element.contains(' ')));
     }
-    return (retVal.length > limit) ? retVal.sublist(0, limit) : retVal;
+    if (retVal.length < limit) {
+      final suggestions = await autoCorrect.suggestionsFor(term, limit);
+      retVal.addAll(suggestions);
+    }
+    final list = retVal.toList();
+    return (list.length > limit) ? list.sublist(0, limit) : list;
   }
 
   @override
-  Future<DictionaryEntry?> getEntry(Term term) => dictionary.getEntry(term);
+  Future<TermProperties?> getEntry(Term term) => dictionary.getEntry(term);
 
   @override
-  Future<Set<String>> synonymsOf(String term) async =>
-      (await getEntry(term))?.allSynonyms ?? {};
+  Future<List<String>> startsWith(String chars, [int limit = 10]) =>
+      autoCorrect.startsWith(chars, limit);
 
   @override
-  Future<List<String>> startsWith(String chars) =>
-      autoCorrect.startsWith(chars);
-
-  @override
-  Future<Map<String, List<String>>> suggestionsFor(String term,
-          [int limit = 10]) =>
+  Future<List<String>> suggestionsFor(String term, [int limit = 10]) =>
       autoCorrect.suggestionsFor(term, limit);
 }
 
@@ -117,9 +127,12 @@ class _DictoSaurusImpl extends DictoSaurusBase {
   final AutoCorrect autoCorrect;
 
   @override
+  Thesaurus get thesaurus => Thesaurus.dictionary(dictionary);
+
+  @override
   final Dictionary dictionary;
 
   /// Initializes a [DictoSaurus] with [autoCorrect], [thesaurus] and
   /// [dictionary] instances.
-  const _DictoSaurusImpl(this.autoCorrect, this.dictionary);
+  const _DictoSaurusImpl(this.dictionary, this.autoCorrect);
 }