+++ lib/src/typedefs.dart
@@ -6,8 +6,9 @@ import 'package:dictosaurus/src/_index.dart';
 
 /// Function definition of an asynchronous callback that returns the meaning of
 /// a [term] from a dictionary provider.
-typedef DictionaryCallback = Future<DictionaryEntry?> Function(String term);
+typedef DictionaryCallback = Future<TermProperties?> Function(String term);
 
 /// Function definition of an asynchronous callback that returns the synonyms of
 /// a [term] from a thesaurus provider.
-typedef SynonymsCallback = Future<Set<String>> Function(String term);
+typedef TermExpander = Future<Set<String>> Function(String term,
+    [PartOfSpeech? partOfSpeech]);