+++ test/dictosaurus_test.dart
@@ -2,42 +2,52 @@
 // BSD 3-Clause License
 // All rights reserved
 
+// ignore_for_file: unused_local_variable
+
 @Timeout(Duration(minutes: 5))
 
 // import the core classes
 import 'package:dictosaurus/dictosaurus.dart';
-// import the typedefs
-import 'package:dictosaurus/type_definitions.dart';
+import '../impl/oxford_dictionary.dart';
+import '../impl/global.dart';
 import 'package:test/test.dart';
-import 'data/sample_news.dart';
-import 'keys/keys.dart';
-import 'impl/oxford_dictionary.dart';
 
 void main() {
   group('Dictionary', (() {
     //
 
     test('Dictionary', (() async {
-      final dictionary = OxfordDictionaries(
-          languageCode: 'en_US',
-          appId: Keys.oxfordDictionaries['appId'] as String,
-          appKey: Keys.oxfordDictionaries['appKey'] as String);
-      // final json = await dictionary.entriesEndPoint('lexicon');
-      final de = await dictionary.getEntry('swim');
-      print(de?.allDefinitions);
+      final term = 'smart';
+      final dictionary = OxfordDictionaries();
+      final definitions = await dictionary.definitionsFor(term);
+      final inflections = await dictionary.inflectionsOf(term);
+      final phrases = await dictionary.phrasesWith(term);
+      print('Thesaurus methods on "$term":');
+      print('Definitions:   $definitions');
+      print('Inflections:   $inflections');
+      print('Phrases:       $phrases');
     }));
   }));
   group('Thesaurus', (() {
     //
 
-    test('Thesaurus()', (() async {}));
+    test('Thesaurus()', (() async {
+      final term = 'swim';
+      final dictionary = OxfordDictionaries();
+      final thesaurus = Thesaurus.dictionary(dictionary);
+      final synonyms = await thesaurus.synonymsOf(term);
+      final antonyms = await thesaurus.antonymsOf(term);
+      print('Thesaurus methods on "$term":');
+      print('Synonyms:      $synonyms');
+      print('Antonyms:      $antonyms');
+    }));
   }));
-  group('Dictosaurus', () {
+  group('Autocorrect', () {
     setUp(() async {});
 
     test('autocorrect', () async {
-      final index = await _getIndex(sampleNews);
-      final autoCorrect = AutoCorrect(index.getKGramIndex, k: 3);
+      final index = await getIndex();
+      final autoCorrect = AutoCorrect(index.getKGramIndex, k: k);
       final term = 'aple';
       final startTime = DateTime.now();
       final start = startTime.millisecondsSinceEpoch;
@@ -53,8 +63,8 @@ void main() {
     });
 
     test('startsWith', () async {
-      final index = await _getIndex(sampleNews);
-      final autoCorrect = AutoCorrect(index.getKGramIndex, k: 3);
+      final index = await getIndex();
+      final autoCorrect = AutoCorrect(index.getKGramIndex, k: k);
       final chars = 'app';
       final startTime = DateTime.now();
       final start = startTime.millisecondsSinceEpoch;
@@ -69,18 +79,80 @@ void main() {
       print('Retrieved $suggestions in $dT seconds.');
     });
   });
-}
+  group('DictoSaurus', () {
+    setUp(() async {});
 
-Future<InvertedIndex> _getIndex(JsonCollection documents,
-    [Map<String, double> zones = const {
-      'name': 1,
-      'descriptions': 0.5
-    }]) async {
-  final index = InMemoryIndex(tokenizer: TextTokenizer(), zones: zones, k: 3);
-  final indexer = TextIndexer(index: index);
-  await indexer.indexCollection(documents);
-  return index;
-}
+    test('DictoSaurus.suggestionsFor("aple")', () async {
+      final dictoSaurus = await getDictoSaurus();
+      final term = 'appel';
+      final startTime = DateTime.now();
+      final start = startTime.millisecondsSinceEpoch;
+      final suggestions = await dictoSaurus.suggestionsFor(term, 10);
 
-Future<Iterable<Token>> tokenizer(String term, [String? zone]) async =>
-    [Token(term, 0)];
+      // get the end time in milliseconds
+      final end = DateTime.now().millisecondsSinceEpoch;
+
+      // calculate the time taken to index the corpus in milliseconds
+      final dT = ((end - start) / 1000).toStringAsFixed(3);
+      // print(suggestions);
+      print('Retrieved $suggestions in $dT seconds.');
+    });
+
+    test('DictoSaurus.startsWith("app")', () async {
+      final dictoSaurus = await getDictoSaurus();
+      final chars = 'app';
+      final startTime = DateTime.now();
+      final start = startTime.millisecondsSinceEpoch;
+      final suggestions = await dictoSaurus.startsWith(chars, 5);
+
+      // get the end time in milliseconds
+      final end = DateTime.now().millisecondsSinceEpoch;
+
+      // calculate the time taken to index the corpus in milliseconds
+      final dT = ((end - start) / 1000).toStringAsFixed(3);
+      // print(suggestions);
+      print('Retrieved $suggestions in $dT seconds.');
+    });
+
+    test('DictoSaurus.expandTerm("apple")', () async {
+      final dictoSaurus = await getDictoSaurus();
+      final term = 'apple';
+      final startTime = DateTime.now();
+      final start = startTime.millisecondsSinceEpoch;
+      final suggestions = await dictoSaurus.expandTerm(term, 25);
+
+      // get the end time in milliseconds
+      final end = DateTime.now().millisecondsSinceEpoch;
+
+      // calculate the time taken to index the corpus in milliseconds
+      final dT = ((end - start) / 1000).toStringAsFixed(3);
+      // print(suggestions);
+      print('Expanded $term to $suggestions in $dT seconds.');
+    });
+
+    test('DictoSaurus.methods', () async {
+      final dictoSaurus = await getDictoSaurus();
+      final misspeltterm = 'inteligent';
+      final startTime = DateTime.now();
+      final start = startTime.millisecondsSinceEpoch;
+      final suggestions = await dictoSaurus.expandTerm(misspeltterm, 25);
+      // get the end time in milliseconds
+      final end = DateTime.now().millisecondsSinceEpoch;
+      // calculate the time taken in milliseconds
+      final dT = ((end - start) / 1000).toStringAsFixed(3);
+      print('Expanded $misspeltterm to $suggestions in $dT seconds.');
+      final term = 'intelligent';
+      final synonyms = await dictoSaurus.synonymsOf(term);
+      final antonyms = await dictoSaurus.antonymsOf(term);
+      final definitions = await dictoSaurus.definitionsFor(term);
+      final inflections = await dictoSaurus.inflectionsOf(term);
+      final phrases = await dictoSaurus.phrasesWith(term);
+      print('Dictosaurus methods on "$term":');
+      print('Definitions:   $definitions');
+      print('Inflections:   $inflections');
+      print('Phrases:       $phrases');
+      print('Synonyms:      $synonyms');
+      print('Antonyms:      $antonyms');
+    });
+  });
+}