+++ CHANGELOG.md
@@ -4,6 +4,20 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+## 0.0.1-beta.7
+
+### *Breaking changes*
+* Renamed `DictionaryEntry` to `TermProperties`.
+* Renamed `TermVariant` to `TermDefinition`.
+* Added `Dictionary.phrasesWith`, `Dictionary.definitionsFor` and `Dictionary.inflectionsOf` methods.
+* Interface `Dictosaurus` implements `Dictionary`, `Thesaurus` and `Autocorrect`.
+
+### *Updated*
+* Documentation.
+* Tests.
+* Examples.
+* Dependencies.
+
 ## 0.0.1-beta.6
 
 ### *Updated*
@@ -14,11 +28,11 @@ All rights reserved.
 **PRE-RELEASE, BREAKING CHANGES**
 
 ### *Breaking changes*
-* Removed [Dictionary.inMemory] factory.
-* Removed [Thesaurus.inMemory] factory.
-* Removed [Dictosaurus.inMemory] factory.
-* Removed [DictionaryMap] typedef.
-* Changed [Dictosaurus] interface.
+* Removed `Dictionary.inMemory` factory.
+* Removed `Thesaurus.inMemory` factory.
+* Removed `Dictosaurus.inMemory` factory.
+* Removed `DictionaryMap` typedef.
+* Changed `Dictosaurus` interface.
 
 ### *Updated*
 * Documentation.
@@ -29,8 +43,8 @@ All rights reserved.
 
 ### *New*
 * Functional early versions of `Dictionary`, `AutoCorrect`, `Thesaurus` and `DictoSaurus`.
-* Object model `DictionaryEntry`.
-* Object model `TermVariant`.
+* Object model `TermProperties`.
+* Object model `TermDefinition`.
 * Enum `PartOfSpeech`.